@@ -8,34 +8,47 @@
 #include <unordered_map>
 #include <string>
 
-class daemon_settings
+class Daemon_settings
 {
 public:
-    typedef int value_t;
+    typedef int int_value_t;
     typedef bool bool_value_t;
 private:
-    std::unordered_map<SettingStringHash::type_t, value_t> settings;
+    std::unordered_map<SettingStringHash::type_t, int_value_t> int_settings;
     std::unordered_map<SettingStringHash::type_t, bool_value_t> bool_settings;
 public:
-    daemon_settings();
+    Daemon_settings();
 
-    value_t get_setting(const std::string& setting_signature) const;
-    void set_setting(const std::string& setting_signature, const value_t& value);
-    bool_value_t get_bool_setting(const std::string& setting_signature) const;
-    void toggle_bool_setting(const std::string& setting_signature);
+    int_value_t get_int_setting(std::string_view setting_signature) const;
+    void set_int_setting(std::string_view setting_signature, const int_value_t& value);
+    bool_value_t get_bool_setting(std::string_view setting_signature) const;
+    void toggle_bool_setting(std::string_view setting_signature);
+
+    bool is_setting(std::string_view setting_signature) const;
+    bool is_bool_setting(std::string_view setting_signature) const;
+    bool is_int_setting(std::string_view setting_signature) const;
+public: // Utility Functions
+    void print_verbose(std::string_view text);
 };
 
-class daemon
+class Daemon
 {
 public:
     typedef std::thread spawn;
 private:
     std::vector<spawn> spawns;
-    daemon_settings settings;
+    std::vector<generate::vector_t*> problem_sets;
+    Daemon_settings settings;
+private:
+    generate::iterator_range_t generate_spawn_range(const generate::problem_count_t& problem_count);
+    bool should_thread(std::string_view process_signature, const generate::problem_count_t& problem_count);
 public:
-    daemon() = default;
+    Daemon() = default;
 
-    void execute(int argc, char** argv);
+    void execute(const std::vector<std::string>& arguments);
 
-    void spawn_process(const std::string& process_signature);
+    void spawn_process(std::string_view process_signature);
+    void cleanup_spawn();
+
+    void ask();
 };
